css: remove the translate -50% hack, and use grid for level icons
Turns out the -50% translate trick was activating sub-pixel positioning
in chrome, and because subpixel positioning apparently work different
for divs than for canvases in this version of chrome, it created gaps
between the canvas and the overlay (the ::before pseudo-element) of the
level icons.

Using the "margin: 0 auto" trick seemed to have fixed that and it all
still works.